Monitor the creation of the Xephyr window and send events to it
While 'Xnest -parent XID' uses the specified window to draw on, the same option
in Xephyr makes it create a subwindow under the specified window.  So, we can no
longer simply forward events to the XID window; we must send them to the subwindow
which Xephyr creates.

So, we catch CreateNotify events under our parent window to see when Xephyr creates
its own subwindow.  We forward events to that window instead.

Signed-off-by: Federico Mena Quintero <federico@novell.com>